+++ core/src/stored/device_resource.h
@@ -72,19 +72,15 @@ class DeviceResource : public BareosResource {
   AutoXflateMode autoinflate; /**< auto inflation in this IO direction */
   utime_t
       vol_poll_interval;   /**< Interval between polling volume during mount */
-  int64_t max_volume_size; /**< Max bytes to put on one volume */
   int64_t max_file_size;   /**< Max file size in bytes */
   int64_t volume_capacity; /**< Advisory capacity */
   int64_t max_spool_size;  /**< Max spool size for all jobs */
   int64_t max_job_spool_size; /**< Max spool size for any single job */
 
-  int64_t max_part_size;    /**< Max part size */
-  char* mount_point;        /**< Mount point for require mount devices */
-  char* mount_command;      /**< Mount command */
-  char* unmount_command;    /**< Unmount command */
-  char* write_part_command; /**< Write part command */
-  char* free_space_command; /**< Free space command */
-  uint32_t count;           /**< Total number of multiplied devices */
+  char* mount_point;     /**< Mount point for require mount devices */
+  char* mount_command;   /**< Mount command */
+  char* unmount_command; /**< Unmount command */
+  uint32_t count;        /**< Total number of multiplied devices */
   DeviceResource* multiplied_device_resource; /**< Copied from this device */
 
   Device* dev; /* Pointer to physical dev -- set at runtime */